+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Added
 ### Fixed
+
+## [5.4.0] - 2024-05-11
+### Fixed
 - Fix a bug where we would re-use the same stream on a failed upload, instead of a new one.
 - Fix a bug where deleting a cloud video would not trigger other clients to update.
 - Make the cloud settings more responsive withb a debounce timer.